backend/requirements.txt
fastapi
uvicorn
websockets
crewai
langchain-openai
langchain-google-genai
pydantic
python-dotenv
